+++ src/pages/_app.tsx
@@ -1,14 +1,18 @@
 import React, { FC } from 'react';
+import { Provider as ReduxProvider } from 'react-redux';
 import { ChakraProvider } from '@chakra-ui/react';
 import type { AppProps } from 'next/app';
 import Layout from '@/components/Layout';
+import store from '@/redux/store';
 
 const App: FC<AppProps> = ({ Component, pageProps }) => (
-  <ChakraProvider>
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
-  </ChakraProvider>
+  <ReduxProvider store={store}>
+    <ChakraProvider>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </ChakraProvider>
+  </ReduxProvider>
 );
 
 export default App;