+++ src/pages/expense.tsx
@@ -1,15 +1,11 @@
 import React, { FC } from 'react';
-import { Box, Heading } from '@chakra-ui/react';
-import ExpenseForm from '@/components/ExpenseForm';
-import ExpenseRadio from '@/components/ExpenseRadio';
+import { Heading } from '@chakra-ui/react';
+import ExpenseCalculator from '@/components/ExpenseCalculator';
 
 const ExpensePage: FC = () => (
   <>
     <Heading>開銷</Heading>
-    <Box w={4 / 5}>
-      <ExpenseRadio />
-    </Box>
-    <ExpenseForm />
+    <ExpenseCalculator />
   </>
 );