+++ src/pages/index.tsx
@@ -1,26 +1,16 @@
 import React, { FC, SyntheticEvent, useState } from 'react';
-import useSWR from 'swr';
 import Link from 'next/link';
-import useUser from '@/hooks/useUser';
-import firebase from '@/utils/firebase';
-import type { AuthResponse } from './api/auth';
-
-const fetcher = (endpoint: string, token: string) =>
-  fetch(`/api/${endpoint}`, {
-    method: 'GET',
-    headers: new Headers({ 'Content-Type': 'application/json', token }),
-    credentials: 'same-origin',
-  }).then((res) => res.json());
+import { useSelector } from 'react-redux';
+import firebase from '@/redux/firebase';
+import { selectUser, logout } from '@/redux/user';
+import { useAppDispatch } from '@/redux/store';
 
 const Home: FC = () => {
+  const { userData } = useSelector(selectUser);
+  const dispatch = useAppDispatch();
+
   const [name, setName] = useState('');
   const [price, setPrice] = useState(0);
-  const { user, logout } = useUser();
-  const { data, error } = useSWR<AuthResponse>(
-    user ? ['auth', user.token] : null,
-    fetcher,
-    { errorRetryCount: 1 },
-  );
 
   const handleOnSubmit = (event: SyntheticEvent): void => {
     event.preventDefault();
@@ -32,7 +22,7 @@ const Home: FC = () => {
     alert('success');
   };
 
-  if (!user) {
+  if (!userData) {
     return (
       <div>
         <p>Hi there!</p>
@@ -49,7 +39,7 @@ const Home: FC = () => {
   return (
     <div>
       <div>
-        <p>You&apos;re signed in. Email: {user.email}</p>
+        <p>You&apos;re signed in. Email: {userData.email}</p>
         <p
           style={{
             display: 'inline-block',
@@ -57,43 +47,35 @@ const Home: FC = () => {
             textDecoration: 'underline',
             cursor: 'pointer',
           }}
-          onClick={() => logout()}
+          onClick={() => dispatch(logout())}
         >
           Log out
         </p>
       </div>
-      {error && <div>Failed to fetch auth status!</div>}
-      {data && !error ? (
-        <>
-          <div>Your status is {data.status}.</div>
-          <form>
-            <div>
-              Name
-              <br />
-              <input
-                type="text"
-                value={name}
-                onChange={(event) => setName(event.target.value)}
-              />
-            </div>
-            <div>
-              Price
-              <br />
-              <input
-                type="number"
-                value={price}
-                min="0"
-                onChange={(event) => setPrice(event.target.valueAsNumber)}
-              />
-            </div>
-            <button type="submit" onClick={handleOnSubmit}>
-              Save
-            </button>
-          </form>
-        </>
-      ) : (
-        <div>Loading...</div>
-      )}
+      <form>
+        <div>
+          Name
+          <br />
+          <input
+            type="text"
+            value={name}
+            onChange={(event) => setName(event.target.value)}
+          />
+        </div>
+        <div>
+          Price
+          <br />
+          <input
+            type="number"
+            value={price}
+            min="0"
+            onChange={(event) => setPrice(event.target.valueAsNumber)}
+          />
+        </div>
+        <button type="submit" onClick={handleOnSubmit}>
+          Save
+        </button>
+      </form>
     </div>
   );
 };