+++ src/redux/collection/selector.ts
@@ -1,0 +1,5 @@
+import type { RootState } from '../store';
+import type { CollectionStates } from './slice';
+
+export const selectCollection = (state: RootState): CollectionStates =>
+  state.collection;