+++ src/redux/reducer.ts
@@ -1,0 +1,9 @@
+import collection from './collection/slice';
+import user from './user/slice';
+
+const reducer = {
+  collection,
+  user,
+};
+
+export default reducer;