+++ src/redux/user/selector.ts
@@ -1,0 +1,4 @@
+import type { RootState } from '../store';
+import type { UserStates } from './slice';
+
+export const selectUser = (state: RootState): UserStates => state.user;